Fix compilation on OS X.
This patch fix several compilation issues on OS X. The most notable
changes are the following.

connectx() is a standard library function on OS X, as a consequence we
must overloaded it using a macro-based trick in order to use our version
of the function.

mremap() is a linux-specific system call. As a consequence, we must use
manually remap the file on other OSes.

Some format strings have been fixed since they did suppose the types of
the parameters to be known, while the standards don't fix them (the most
notable one being struct timeval.tv_usec).

+++ http-srv-static.c
@@ -183,10 +183,10 @@ void httpd_reply_file(httpd_query_t *q, int dfd, const char *file, bool head)
     httpd_put_date_hdr(ob, "Last-Modified", st.st_mtime);
     if (st.st_mtime >= lp_getsec() - 10) {
         ob_addf(ob, "ETag: W/\"%jx-%jxx-%lx\"\r\n",
-                st.st_ino, st.st_size, st.st_mtime);
+                (int64_t)st.st_ino, st.st_size, st.st_mtime);
     } else {
         ob_addf(ob, "ETag: \"%jx-%jxx-%lx\"\r\n",
-                st.st_ino, st.st_size, st.st_mtime);
+                (int64_t)st.st_ino, st.st_size, st.st_mtime);
     }
     mime_put_http_ctype(ob, file);
     httpd_reply_hdrs_done(q, st.st_size, false);
